requirements.txt
alembic==0.8.2
Flask==0.10.1
Flask-Script==2.0.5
Flask-Migrate==1.6.0
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.4
cookies==2.2.1
coverage==3.7.1
funcsigs==0.4
honcho==0.6.6
itsdangerous==0.24
libsass==0.8.3
mock==1.1.2
nose==1.3.7
pbr==1.3.0
requests==2.7.0
requests-mock==0.6.0
responses==0.4.0
six==1.9.0
waitress==0.8.9
psycopg2==2.6.1
google-api-python-client
